requirements.txt
# Core ML and optimization libraries
torch>=1.12.0
botorch>=0.8.0
gpytorch>=1.9.0

# Data manipulation and analysis
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization and Explainability
shap>=0.40.0
matplotlib>=3.5.0
seaborn>=0.11.0

# File I/O
openpyxl>=3.0.0
pyyaml>=6.0
joblib>=1.1.0
